% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_network_format.R
\name{change_network_format}
\alias{change_network_format}
\title{Change Network Format}
\usage{
change_network_format(net, ng = NULL)
}
\arguments{
\item{net}{matrix containing the network in the from of adjaceny matrix or edgelist}

\item{ng}{Total number of genes of the network. Defaults to NULL.}
}
\description{
This function changes a network from an adjecy matrix to an edgelist 
matrix and viceversa
}
\examples{
net <- matrix(rnorm(100),10,10)
Edge_list <- change_network_format(net)
net2 <- change_network_format(Edge_list)
}
