% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DREAM_Evaluation.R
\name{DREAM_Evaluation}
\alias{DREAM_Evaluation}
\title{DREAM GRN Evaluation}
\usage{
DREAM_Evaluation(gold_positives, prediction_raw)
}
\arguments{
\item{gold_positives}{matrix with three columns that contains the GS in the form 
of edgelist.}

\item{predict_raw}{matrix with three columns that contains the inferred GRN in the form 
of edgelist}
}
\value{
list containing the metrics [TPR FPR PREC REC L AUROC AUPR] 
TPR: vector with True Positive Ratio
FPR: vector with False Positive Ratio
PREC: vector with Precision
REC: vector with Recall
L: prediction list length
AUROC: area under ROC
AUPR: area under PR
}
\description{
This function allows you to evaluate a GRN inference against golden standard (GS).
}
\examples{
net <- matrix(rnorm(100),10,10)
GS <- matrix(rnorm(100),10,10)
GS[which(GS>0.5)]<-1
GS[GS!=1] <- 0
diag(GS) <- 0
net <- change_network_format(net)
GS <- change_network_format(GS)
Ev <- DREAM_Evaluation(GS,net)
}
\keyword{main}
