% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure.R
\name{Figure}
\alias{Figure}
\title{Plot ROC or PR curve}
\usage{
figure(ev, fig = "pr", two = FALSE, ...)
}
\arguments{
\item{ev}{list containg the evaluation' metrics. Output of DREAM_Evaluation function}

\item{fig}{charater of the kind of plot ("pr" or "roc")}

\item{two}{logical indicating if the random extension should be marked or not}

\item{...}{additional parameter to the plot}
}
\description{
Plot ROC or PR curve
}
\examples{
net <- matrix(rnorm(100),10,10)
GS <- matrix(rnorm(100),10,10)
GS[which(GS>0.5)]<-1
GS[GS!=1] <- 0
diag(GS) <- 0
net <- change_network_format(net)
GS <- change_network_format(GS)
Ev <- DREAM_Evaluation(GS,net)
Figure(Ev)
}
