% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadGS.R
\name{LoadGS}
\alias{LoadGS}
\title{Load Golden standard data}
\usage{
LoadGS(net = 1)
}
\arguments{
\item{net}{integer identifier of the network}
}
\value{
l
}
\description{
Load Golden standard data
}
\examples{
Data <- LaodGS(1)
}
